+++ backend/policy-tracker/src/main/java/com/tunahancoban/policy_tracker/model/DTO/response/CustomerSummaryResponse.java
@@ -6,12 +6,14 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 public class CustomerSummaryResponse {
-    private long totalPremium;
+    private BigDecimal totalPremium;
     private long activePolicyNumber;
     private long expiringSoonPolicies;
     private long expiredPolicies;
